requirements.txt
PyRSS2Gen
beautifulsoup4 # BeautifulSoup
Mastodon.py
feedparser
